🐛 fix: node repository 삭제

+++ be/smartfactory/src/main/java/info/smartfactory/domain/bottleneck/dto/request/AddBottleneckRequest.java
@@ -16,7 +16,7 @@ public class AddBottleneckRequest {
     @NotNull(message = "y좌표는 필수 항목입니다.")
     int y_coordinate;
 
-    int missionId;
+    Long missionId;
 
     LocalDateTime bottleneckCreatedAt;
 }
